add codeM_2017

+++ codeCompete/src/CodeM/Digital_2017.java
@@ -8,16 +8,16 @@ public class Digital_2017 {
         //System.out.println(getFirst(980));
         Scanner sc=new Scanner(System.in);
         while (sc.hasNext()){
-         int left=sc.nextInt();
-         int right=sc.nextInt();
-         int[] res=new int[10];
+         long left=sc.nextLong();
+         long right=sc.nextLong();
+         long[] res=new long[10];
          //res[1]+=right-left+1;
             int t;
          int end=(int) Math.sqrt(right);
          for (int k=1;k<=end;k++){
              int cnt=0;
-             int max=Math.max(left,k);
-             for (int i=max;i<=right;i++){
+             long max=Math.max(left,k);
+             for (long i=max;i<=right;i++){
                  if (i%k==0) {
                      cnt += 1;
                      if (i / k > end) {
@@ -33,13 +33,13 @@ public class Digital_2017 {
              System.out.println(res[t]);
         }
     }
-    private static int  getFirst(int k){
-        int res=0;
+    private static int getFirst(long k){
+        long res=0;
         while (k!=0){
             res=k;
             k/=10;
         }
-        return res;
+        return (int)res;
     }
     private static int getFirst1(int k){
         //System.out.println(k);
